+++ packages/client/src/components/app/forms/Field.svelte
@@ -6,9 +6,8 @@
   import { memo } from "@budibase/frontend-core"
   import Placeholder from "../Placeholder.svelte"
   import InnerForm from "./InnerForm.svelte"
-  import type { FieldApi, FieldState } from "."
   import type { FieldSchema, FieldType } from "@budibase/types"
-  import type { FieldValidation, FormField } from "@/index"
+  import { FieldApi, FieldState, FieldValidation, FormField } from "@/types"
 
   interface FieldInfo {
     field: string
