+++ packages/client/src/types/components.ts
@@ -1,0 +1,9 @@
+import { Readable } from "svelte/store"
+
+export type Component = Readable<{
+  id: string
+  name: string
+  styles: any
+  editing: boolean
+  errorState: boolean
+}>
