+++ packages/client/src/types/fields.ts
@@ -1,0 +1,3 @@
+export interface FieldGroupContext {
+  labelPosition: string
+}
